Clean-up Swashbuckle (#24150)
