+++ aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/2.0/TodoApi.NSwag/Controllers/TodoController.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using TodoApi.Models;
 
-#region snippet_TodoController
+// <snippet_TodoController>
 namespace TodoApi.Controllers
 {
     [Produces("application/json")]
@@ -12,7 +12,7 @@ namespace TodoApi.Controllers
     public class TodoController : Controller
     {
         private readonly TodoContext _context;
-        #endregion snippet_TodoController
+        // </snippet_TodoController>
 
         public TodoController(TodoContext context)
         {
@@ -25,16 +25,16 @@ namespace TodoApi.Controllers
             }
         }
 
-        #region snippet_GetAll
+        // <snippet_GetAll>
         [HttpGet]
         [Produces("application/json", Type = typeof(TodoItem))]
         public List<TodoItem> GetAll()
         {
             return _context.TodoItems.ToList();
         }
-        #endregion snippet_GetAll
+        // </snippet_GetAll>
 
-        #region snippet_GetById
+        // <snippet_GetById>
         [HttpGet("{id}", Name = "GetTodo")]
         [Produces("application/json", Type = typeof(TodoItem))]
         public IActionResult GetById(long id)
@@ -48,9 +48,9 @@ namespace TodoApi.Controllers
 
             return Ok(item);
         }
-        #endregion snippet_GetById
+        // </snippet_GetById>
 
-        #region snippet_Create
+        // <snippet_Create>
         /// <summary>
         /// Creates a TodoItem.
         /// </summary>
@@ -69,11 +69,11 @@ namespace TodoApi.Controllers
         /// <returns>A newly created TodoItem</returns>
         /// <response code="201">Returns the newly created item</response>
         /// <response code="400">If the item is null</response>
-        #region snippet_CreateActionAttributes
+        // <snippet_CreateActionAttributes>
         [ProducesResponseType(typeof(TodoItem), StatusCodes.Status201Created)]  // Created
         [ProducesResponseType(StatusCodes.Status400BadRequest)]                 // BadRequest
-        #endregion snippet_CreateActionAttributes
-        #region snippet_CreateAction
+        // </snippet_CreateActionAttributes>
+        // <snippet_CreateAction>
         [HttpPost]
         public IActionResult Create([FromBody] TodoItem item)
         {
@@ -87,10 +87,10 @@ namespace TodoApi.Controllers
 
             return CreatedAtRoute("GetTodo", new { id = item.Id }, item);
         }
-        #endregion snippet_CreateAction
-        #endregion snippet_Create
+        // </snippet_CreateAction>
+        // </snippet_Create>
 
-        #region snippet_Update
+        // <snippet_Update>
         [HttpPut("{id}")]
         public IActionResult Update(long id, [FromBody] TodoItem item)
         {
@@ -113,9 +113,9 @@ namespace TodoApi.Controllers
 
             return NoContent();
         }
-        #endregion snippet_Update
+        // </snippet_Update>
 
-        #region snippet_Delete
+        // <snippet_Delete>
         /// <summary>
         /// Deletes a specific TodoItem.
         /// </summary>
@@ -133,6 +133,6 @@ namespace TodoApi.Controllers
             _context.SaveChanges();
             return NoContent();
         }
-        #endregion snippet_Delete
+        // </snippet_Delete>
     }
 }
