+++ aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/2.0/TodoApi.NSwag/Startup.cs
@@ -7,7 +7,7 @@ namespace TodoApi
 {
     public class Startup
     {
-        #region snippet_ConfigureServices
+        // <snippet_ConfigureServices>
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddDbContext<TodoContext>(opt =>
@@ -17,9 +17,9 @@ namespace TodoApi
             // Register the Swagger services
             services.AddSwaggerDocument();
         }
-        #endregion snippet_ConfigureServices
+        // </snippet_ConfigureServices>
 
-        #region snippet_Configure
+        // <snippet_Configure>
         public void Configure(IApplicationBuilder app)
         {
             app.UseStaticFiles();
@@ -30,6 +30,6 @@ namespace TodoApi
 
             app.UseMvc();
         }
-        #endregion snippet_Configure
+        // </snippet_Configure>
     }
 }
