+++ aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/2.0/TodoApi.Swashbuckle/Startup2.cs
@@ -1,20 +1,20 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-#region snippet_InfoClassNamespace
+// <snippet_InfoClassNamespace>
 using Microsoft.OpenApi.Models;
-#endregion
+// </snippet_InfoClassNamespace>
 using TodoApi.Models;
-#region snippet_PreReqNamespaces
+// <snippet_PreReqNamespaces>
 using System;
 using System.Reflection;
 using System.IO;
-#endregion
+// </snippet_PreReqNamespaces>
 namespace TodoApi
 {
     public class Startup2
     {
-        #region snippet_ConfigureServices
+        // <snippet_ConfigureServices>
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddDbContext<TodoContext>(opt =>
@@ -24,9 +24,9 @@ namespace TodoApi
             // Register the Swagger generator, defining 1 or more Swagger documents
             services.AddSwaggerGen();
         }
-        #endregion
+        // </snippet_ConfigureServices>
 
-        #region snippet_Configure
+        // <snippet_Configure>
         public void Configure(IApplicationBuilder app)
         {
             // Enable middleware to serve generated Swagger as a JSON endpoint.
@@ -41,6 +41,6 @@ namespace TodoApi
 
             app.UseMvc();
         }
-        #endregion
+        // </snippet_Configure>
     }
 }
