+++ aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/2.1/TodoApi.NSwag/Controllers/TodoController.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using TodoApi.Models;
 
-#region snippet_TodoController
+// <snippet_TodoController>
 namespace TodoApi.Controllers
 {
     [Produces("application/json")]
@@ -13,7 +13,7 @@ namespace TodoApi.Controllers
     public class TodoController : Controller
     {
         private readonly TodoContext _context;
-        #endregion snippet_TodoController
+        // </snippet_TodoController>
 
         public TodoController(TodoContext context)
         {
@@ -26,15 +26,15 @@ namespace TodoApi.Controllers
             }
         }
 
-        #region snippet_GetAll
+        // <snippet_GetAll>
         [HttpGet]
         public ActionResult<List<TodoItem>> GetAll()
         {
             return _context.TodoItems.ToList();
         }
-        #endregion snippet_GetAll
+        // </snippet_GetAll>
 
-        #region snippet_GetById
+        // <snippet_GetById>
         [HttpGet("{id}", Name = "GetTodo")]
         public ActionResult<TodoItem> GetById(long id)
         {
@@ -47,9 +47,9 @@ namespace TodoApi.Controllers
 
             return item;
         }
-        #endregion snippet_GetById
+        // </snippet_GetById>
 
-        #region snippet_Create
+        // <snippet_Create>
         /// <summary>
         /// Creates a TodoItem.
         /// </summary>
@@ -68,11 +68,11 @@ namespace TodoApi.Controllers
         /// <returns>A newly created TodoItem</returns>
         /// <response code="201">Returns the newly created item</response>
         /// <response code="400">If the item is null</response>
-        #region snippet_CreateActionAttributes
+        // <snippet_CreateActionAttributes>
         [ProducesResponseType(StatusCodes.Status201Created)]     // Created
         [ProducesResponseType(StatusCodes.Status400BadRequest)]  // BadRequest
-        #endregion snippet_CreateActionAttributes
-        #region snippet_CreateAction
+        // </snippet_CreateActionAttributes>
+        // <snippet_CreateAction>
         [HttpPost]
         public ActionResult<TodoItem> Create(TodoItem item)
         {
@@ -81,10 +81,10 @@ namespace TodoApi.Controllers
 
             return CreatedAtRoute("GetTodo", new { id = item.Id }, item);
         }
-        #endregion snippet_CreateAction
-        #endregion snippet_Create
+        // </snippet_CreateAction>
+        // </snippet_Create>
 
-        #region snippet_Update
+        // <snippet_Update>
         [HttpPut("{id}")]
         public IActionResult Update(long id, TodoItem item)
         {
@@ -107,9 +107,9 @@ namespace TodoApi.Controllers
 
             return NoContent();
         }
-        #endregion snippet_Update
+        // </snippet_Update>
 
-        #region snippet_Delete
+        // <snippet_Delete>
         /// <summary>
         /// Deletes a specific TodoItem.
         /// </summary>
@@ -127,6 +127,6 @@ namespace TodoApi.Controllers
             _context.SaveChanges();
             return NoContent();
         }
-        #endregion snippet_Delete
+        // </snippet_Delete>
     }
 }
