+++ aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/3.x/TodoApi.Swashbuckle/Startup.cs
@@ -12,7 +12,7 @@ namespace TodoApi
 {
     public class Startup
     {
-        #region snippet_ConfigureServices
+        // <snippet_ConfigureServices>
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddDbContext<TodoContext>(opt =>
@@ -47,9 +47,9 @@ namespace TodoApi
                 c.IncludeXmlComments(xmlPath);
             });
         }
-        #endregion
+        // </snippet_ConfigureServices>
 
-        #region snippet_Configure
+        // <snippet_Configure>
         public void Configure(IApplicationBuilder app)
         {
             app.UseStaticFiles();
@@ -70,6 +70,6 @@ namespace TodoApi
                 endpoints.MapControllers();
             });
         }
-        #endregion
+        // </snippet_Configure>
     }
 }
