+++ aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/3.x/TodoApi.Swashbuckle/sample_images/CreateLocalCopy-NSwagStudio.PNG
