--- aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/6.0/SwashbuckleSample/Models/TodoItem.cs
@@ -1,16 +1,0 @@
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
-
-namespace SwashbuckleSample.Models
-{
-    public class TodoItem
-    {
-        public long Id { get; set; }
-
-        [Required]
-        public string Name { get; set; } = null!;
-
-        [DefaultValue(false)]
-        public bool IsComplete { get; set; }
-    }
-}
