+++ aspnetcore/tutorials/web-api-help-pages-using-swagger/samples/6.x/SwashbuckleSample/appsettings.Development.json
